pkg/godax/market_data: add support for fetching a product’s order book
This includes a custom JSON unmarshaller because coinbase returns
JSON with “bids” and “asks” which both are arrays with multiple types
(strings and ints). Found a couple inconsistencies in the API which I
noted in some TODOs. One of the errors is when you ask for “level=3”
in the query param, those “bids” and “asks” arrays no longer have
what look like NumOrders as an int, but instead a UUID string of sorts.